@@ -41,42 +41,42 @@ class LengthType extends AbstractType
     {
         $resolver->setDefaults([
             'choices' => [
-                Ban::LENGTH_PERMANENT => 'Permanent',
+                'Permanent' => Ban::LENGTH_PERMANENT,
                 'minutes' => [
-                  1  => $this->transChoice('{n} minute|{n} minutes', 1),
-                  5  => $this->transChoice('{n} minute|{n} minutes', 5),
-                  10 => $this->transChoice('{n} minute|{n} minutes', 10),
-                  15 => $this->transChoice('{n} minute|{n} minutes', 15),
-                  30 => $this->transChoice('{n} minute|{n} minutes', 30),
-                  45 => $this->transChoice('{n} minute|{n} minutes', 45),
+                  $this->transChoice('{n} minute|{n} minutes', 1)  => 1,
+                  $this->transChoice('{n} minute|{n} minutes', 5)  => 5,
+                  $this->transChoice('{n} minute|{n} minutes', 10) => 10,
+                  $this->transChoice('{n} minute|{n} minutes', 15) => 15,
+                  $this->transChoice('{n} minute|{n} minutes', 30) => 30,
+                  $this->transChoice('{n} minute|{n} minutes', 45) => 45,
                 ],
                 'hours' => [
-                  60  => $this->transChoice('{n} hour|{n} hours', 1),
-                  120 => $this->transChoice('{n} hour|{n} hours', 2),
-                  180 => $this->transChoice('{n} hour|{n} hours', 3),
-                  240 => $this->transChoice('{n} hour|{n} hours', 4),
-                  480 => $this->transChoice('{n} hour|{n} hours', 8),
-                  720 => $this->transChoice('{n} hour|{n} hours', 12),
+                  $this->transChoice('{n} hour|{n} hours', 1)  => 60,
+                  $this->transChoice('{n} hour|{n} hours', 2)  => 120,
+                  $this->transChoice('{n} hour|{n} hours', 3)  => 180,
+                  $this->transChoice('{n} hour|{n} hours', 4)  => 240,
+                  $this->transChoice('{n} hour|{n} hours', 8)  => 480,
+                  $this->transChoice('{n} hour|{n} hours', 12) => 720,
                 ],
                 'days' => [
-                  1440 => $this->transChoice('{n} day|{n} days', 1),
-                  2880 => $this->transChoice('{n} day|{n} days', 2),
-                  4320 => $this->transChoice('{n} day|{n} days', 3),
-                  5760 => $this->transChoice('{n} day|{n} days', 4),
-                  7200 => $this->transChoice('{n} day|{n} days', 5),
-                  8640 => $this->transChoice('{n} day|{n} days', 6),
+                  $this->transChoice('{n} day|{n} days', 1) => 1440,
+                  $this->transChoice('{n} day|{n} days', 2) => 2880,
+                  $this->transChoice('{n} day|{n} days', 3) => 4320,
+                  $this->transChoice('{n} day|{n} days', 4) => 5760,
+                  $this->transChoice('{n} day|{n} days', 5) => 7200,
+                  $this->transChoice('{n} day|{n} days', 6) => 8640,
                 ],
                 'weeks' => [
-                  10080 => $this->transChoice('{n} week|{n} weeks', 1),
-                  20160 => $this->transChoice('{n} week|{n} weeks', 2),
-                  30240 => $this->transChoice('{n} week|{n} weeks', 3),
+                  $this->transChoice('{n} week|{n} weeks', 1) => 10080,
+                  $this->transChoice('{n} week|{n} weeks', 2) => 20160,
+                  $this->transChoice('{n} week|{n} weeks', 3) => 30240,
                 ],
                 'months' => [
-                  43200  => $this->transChoice('{n} month|{n} months', 1),
-                  86400  => $this->transChoice('{n} month|{n} months', 2),
-                  129600 => $this->transChoice('{n} month|{n} months', 3),
-                  259200 => $this->transChoice('{n} month|{n} months', 6),
-                  518400 => $this->transChoice('{n} month|{n} months', 12),
+                  $this->transChoice('{n} month|{n} months', 1)  => 43200,
+                  $this->transChoice('{n} month|{n} months', 2)  => 86400,
+                  $this->transChoice('{n} month|{n} months', 3)  => 129600,
+                  $this->transChoice('{n} month|{n} months', 6)  => 259200,
+                  $this->transChoice('{n} month|{n} months', 12) => 518400,
                 ],
             ],
         ]);
