@@ -27,27 +27,27 @@ class FlagType extends AbstractType
     {
         $resolver->setDefaults([
             'choices' => [
-                Admin::FLAG_ROOT        => 'flags.root',
-                Admin::FLAG_RESERVATION => 'flags.reservation',
-                Admin::FLAG_GENERIC     => 'flags.generic',
-                Admin::FLAG_KICK        => 'flags.kick',
-                Admin::FLAG_BAN         => 'flags.ban',
-                Admin::FLAG_UNBAN       => 'flags.unban',
-                Admin::FLAG_SLAY        => 'flags.slay',
-                Admin::FLAG_CHANGEMAP   => 'flags.changemap',
-                Admin::FLAG_CONVARS     => 'flags.convars',
-                Admin::FLAG_CONFIG      => 'flags.config',
-                Admin::FLAG_CHAT        => 'flags.chat',
-                Admin::FLAG_VOTE        => 'flags.vote',
-                Admin::FLAG_PASSWORD    => 'flags.password',
-                Admin::FLAG_RCON        => 'flags.rcon',
-                Admin::FLAG_CHEATS      => 'flags.cheats',
-                Admin::FLAG_CUSTOM1     => 'flags.custom1',
-                Admin::FLAG_CUSTOM2     => 'flags.custom2',
-                Admin::FLAG_CUSTOM3     => 'flags.custom3',
-                Admin::FLAG_CUSTOM4     => 'flags.custom4',
-                Admin::FLAG_CUSTOM5     => 'flags.custom5',
-                Admin::FLAG_CUSTOM6     => 'flags.custom6',
+                'flags.root'        => Admin::FLAG_ROOT,
+                'flags.reservation' => Admin::FLAG_RESERVATION,
+                'flags.generic'     => Admin::FLAG_GENERIC,
+                'flags.kick'        => Admin::FLAG_KICK,
+                'flags.ban'         => Admin::FLAG_BAN,
+                'flags.unban'       => Admin::FLAG_UNBAN,
+                'flags.slay'        => Admin::FLAG_SLAY,
+                'flags.changemap'   => Admin::FLAG_CHANGEMAP,
+                'flags.convars'     => Admin::FLAG_CONVARS,
+                'flags.config'      => Admin::FLAG_CONFIG,
+                'flags.chat'        => Admin::FLAG_CHAT,
+                'flags.vote'        => Admin::FLAG_VOTE,
+                'flags.password'    => Admin::FLAG_PASSWORD,
+                'flags.rcon'        => Admin::FLAG_RCON,
+                'flags.cheats'      => Admin::FLAG_CHEATS,
+                'flags.custom1'     => Admin::FLAG_CUSTOM1,
+                'flags.custom2'     => Admin::FLAG_CUSTOM2,
+                'flags.custom3'     => Admin::FLAG_CUSTOM3,
+                'flags.custom4'     => Admin::FLAG_CUSTOM4,
+                'flags.custom5'     => Admin::FLAG_CUSTOM5,
+                'flags.custom6'     => Admin::FLAG_CUSTOM6,
             ],
             'expanded' => true,
             'multiple' => true,
